Get well names from top left and bottom right wells

Get well names from top left and bottom right wells

Usage

get_wells_from_edges(
  top_left,
  bottom_right,
  return_type = c("wells", "rows_and_columns")
)

See also