Line plot of clone read fraction across multiple samples

Line plot of clone read fraction across multiple samples

Usage

plot_clone_size_across_samples(
  data,
  clones,
  chain = c("beta", "alpha"),
  pseudo = 1e-06,
  group_vec = NULL,
  sum_readFraction = TRUE,
  samples = NULL,
  return_data = FALSE,
  label_zero = FALSE,
  show_legend = TRUE
)

Arguments

data

a TIRTLseqData object

clones

a list of nucleotide sequences of TCR clones

chain

the TCR chain used (default is "beta")

pseudo

the value of the pseudocount added to all read fractions (default is 10^-6)

group_vec

(optional) a vector of "groups" for the clones

sum_readFraction

if TRUE, plot the sum of read fractions of clones in each group. If FALSE, plot each clone with a separate line.

samples

(optional) which samples to use in the plot (default is all samples)

return_data

whether to return the data used for plotting (default is FALSE)

label_zero

whether to label zero on the y-axis (default is FALSE)

show_legend

whether to show the legend (default is TRUE)

See also