Plot the overlap/agreement between samples (in terms of most frequent clones)

plot_sample_overlap() returns a heatmap showing the overlap in most frequent TCRs among pairs of samples in a dataset

Usage

plot_sample_overlap(
  data,
  chain = c("paired", "alpha", "beta"),
  n_seq = 200,
  show_row_names = FALSE,
  show_column_names = FALSE,
  label_col = "Sample",
  title = "",
  return_data = FALSE,
  cluster_rows = TRUE,
  cluster_cols = TRUE
)

sample_overlap(...)

Arguments

data

the dataset, an object loaded using the load_tirtlseq() function

chain

which chain to plot: either paired or alpha-/beta-pseudobulk. (default "paired")

n_seq

the number of most frequent TCR sequences to use (default 200)

show_row_names

whether to show row names for the heatmap (default FALSE)

show_column_names

whether to show column names for the heatmap (default FALSE)

label_col

a column of the metadata to use as labels for rows and columns (default "Sample", uses the sample_id)

title

a title for the heatmap

Value

A heatmap with hierarchically clustered rows and columns showing the number of TCRs shared between each pair of samples, among their top N most frequent TCRs.

Details

The function takes the top N most frequent TCRs found in each dataset (default 200) and compares their overlap between samples.

See also

Examples

# example code
# data = load_tirtlseq("your_directory/")
# plot_sample_overlap(data, chain = "paired", n_seq=200)