Create a summary table with number of reads and unique alpha/beta chains observed for each sample

Create a summary table with number of reads and unique alpha/beta chains observed for each sample

Usage

summarize_data(data)

Arguments

data

a TIRTLseqData object

See also